Renamed WindowGLFW to GlfwWindow
Closes https://github.com/PistonDevelopers/glfw_window/issues/53
